refactor: #684 NavigationEditor 책임 분리
- src/components/shared/admin/navigation/ 디렉토리로 하위 컴포넌트 분리
  - navigationGroups.ts: GROUP_INFO 상수 + NavGroup 타입
  - flattenItems.ts: 트리 평탄화 유틸
  - NavigationPreview.tsx: GNB/사이드바/푸터 미리보기 (+ GNBDropdownPreviewItem)
  - SortableNavigationRow.tsx: 드래그 가능한 재귀 트리 행 + 액션 버튼
  - NavigationFormModal.tsx: 메뉴 추가/수정 폼 모달
- NavigationEditor.tsx 는 에디터 셸 (상태·DnD 컨텍스트·액션바) 로 축소 (607 → 192 lines)
- 공개 API (Props) 변경 없음 → /admin/navigation 페이지 수정 불필요

Closes #684
